refactor(di): move all binding resolution logic into injector.js
